Appearance/Some functionality
I made the simple style of the page, its pretty basic because my main focus is just to build a todo list and learn as much as i can from the project. So far all the list can do is list the stuff written in the input text after you click the create button.